agent-examples/03-pydanticai-structured-agent/requirements.txt
python-dotenv
pydantic
pydantic-ai
